code cleanup: useless casts
git-svn-id: https://svn.code.sf.net/p/itext/code/trunk@3427 3a709846-db2a-4fbd-be0b-eb932eee6479

+++ itext/src/core/com/lowagie/text/pdf/BarcodePDF417.java
@@ -483,7 +483,7 @@ public class BarcodePDF417 {
             for (k = retLast; k >= 0; --k)
                 codewords[ret + k] *= 256;
             // add the digit
-            codewords[ret + retLast] += (int)text[ni] & 0xff;
+            codewords[ret + retLast] += text[ni] & 0xff;
             // propagate carry
             for (k = retLast; k > 0; --k) {
                 codewords[ret + k - 1] += codewords[ret + k] / 900;
@@ -503,7 +503,7 @@ public class BarcodePDF417 {
             size = length - k < 44 ? length - k : 6;
             if (size < 6) {
                 for (j = 0; j < size; ++j)
-                    codewords[cwPtr++] = (int)text[k + j] & 0xff;
+                    codewords[cwPtr++] = text[k + j] & 0xff;
             }
             else {
                 byteCompaction6(k);
